Avoid managing build user group on Windows platforms
Commit feec1ef7 introduced changes to ignore failures around creating groups on
Windows platforms. This change seeks to extend that premise to other resources in
the Chef run which depend on the changes in feec1ef7.

When the initial creation of the omnibus `group` resource fails, and the failure
is ignored, subsequently attempting to append the omnibus user to that group will
also fail and should be ignored.

When both instances of the group resource configured with `ignore_failure true if
windows?`, we must also exclude the `group` property from `directory` resources
when executing on windows. Without excluding the group property

This change was necessary for our team to successfully build our omnibus project
on Windows 2012r2 images in Amazon EC2. h/t @atomic-penguin for PR #189 which
put us on the right track.

Closes #189

Signed-off-by: Cameron Johnston <cameron@rootdown.net>